chore: update to 0.32

+++ decidim-restfull-core/lib/decidim/rest_full/core/http_cache/collection_fingerprint.rb
@@ -68,7 +68,7 @@ module Decidim
             input = [
               @ctx.organization.id,
               @ctx.profile,
-              last_modified.to_i,
+              last_modified.to_f,
               count,
               @ctx.client_id,
               @ctx.act_as&.id,
